requirements.txt
adapt==0.4.4
black==24.10.0
flake8==7.1.1
isort==5.13.2
lightgbm==4.5.0
matplotlib==3.9.2
mypy==1.13.0
numpy==2.1.0
optuna==4.2.0
pandas==2.2.2
pre-commit==4.0.1
pyyaml==6.0.2
scikit-learn==1.5.2
torch==2.4.0
